Avoid signed integer overflow in example sketch
The `int' type is only 16-bits long in some Arduinos. On those devices,
the square of an analog reading, filtered or not, may not fit in an
`int'. Use `long' instead, in order to avoid signed integer overflow,
which in C++ is undefined behavior.

+++ examples/SimpleEMGFilters/SimpleEMGFilters.ino
@@ -98,10 +98,10 @@ void loop() {
     int data = analogRead(SensorInputPin);
 
     // filter processing
-    int dataAfterFilter = myFilter.update(data);
+    long dataAfterFilter = myFilter.update(data);
 
     // Get envelope by squaring the input
-    int envelope = sq(dataAfterFilter);
+    long envelope = sq(dataAfterFilter);
 
     if (CALIBRATE) {
         Serial.print("Squared Data: ");
